+++ threedi_models_and_simulations/utils.py
@@ -23,11 +23,13 @@ EMPTY_DB_PATH = os.path.join(DATA_PATH, "empty.sqlite")
 CHUNK_SIZE = 1024**2
 RADAR_ID = "d6c2347d-7bd1-4d9d-a1f6-b342c865516f"
 API_DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%f%z"
+USER_DATETIME_FORMAT = "%Y-%m-%d %H:%M:%S"
 
 
 class EventTypes(Enum):
     CONSTANT = "Constant"
-    CUSTOM = "Custom"
+    FROM_CSV = "From CSV"
+    FROM_NETCDF = "From NetCDF"
     DESIGN = "Design"
     RADAR = "Radar - NL Only"