+++ threedi_models_and_simulations/workers.py
@@ -67,6 +67,11 @@ class WSProgressesSentinel(QObject):
     @pyqtSlot()
     def run(self):
         """Checking running simulations progresses."""
+        self.fetch_finished_simulations()
+        self.start_listening()
+
+    def fetch_finished_simulations(self):
+        """Fetches finished simulations data."""
         try:
             self.tc = ThreediCalls(self.threedi_api)
             logger.debug("Fetching finished simulation statuses")
@@ -92,11 +97,6 @@ class WSProgressesSentinel(QObject):
         except ApiException as e:
             error_msg = extract_error_message(e)
             self.thread_failed.emit(error_msg)
-            return
-        self.ws_client = QtWebSockets.QWebSocket(version=QtWebSockets.QWebSocketProtocol.VersionLatest)
-        self.ws_client.textMessageReceived.connect(self.all_simulations_progress_web_socket)
-        self.ws_client.error.connect(self.websocket_error)
-        self.start_listening()
 
     def start_listening(self):
         """Start listening of active simulations websocket."""
@@ -106,16 +106,20 @@ class WSProgressesSentinel(QObject):
         api_version = self.tc.threedi_api.version
         ws_request = QNetworkRequest(QUrl(f"{self.wss_url}/{api_version}/active-simulations/"))
         ws_request.setRawHeader(QByteArray().append("Authorization"), QByteArray().append(basic_auth_token))
+        self.ws_client = QtWebSockets.QWebSocket(version=QtWebSockets.QWebSocketProtocol.VersionLatest)
+        self.ws_client.textMessageReceived.connect(self.all_simulations_progress_web_socket)
+        self.ws_client.error.connect(self.websocket_error)
         self.ws_client.open(ws_request)
 
-    def stop_listening(self):
+    def stop_listening(self, be_quite=False):
         """Close websocket client."""
         if self.ws_client is not None:
             self.ws_client.textMessageReceived.disconnect(self.all_simulations_progress_web_socket)
             self.ws_client.error.disconnect(self.websocket_error)
             self.ws_client.close()
-            stop_message = "Checking running simulation stopped."
-            self.thread_finished.emit(stop_message)
+            if be_quite is False:
+                stop_message = "Checking running simulation stopped."
+                self.thread_finished.emit(stop_message)
 
     def websocket_error(self, error_code):
         """Report errors from websocket."""
@@ -913,30 +917,37 @@ class SimulationRunner(QRunnable):
             offset = self.current_simulation.precipitation.offset
             start = self.current_simulation.precipitation.start
             interpolate = self.current_simulation.precipitation.interpolate
-            filepath = self.current_simulation.precipitation.filepath
-            from_csv = self.current_simulation.precipitation.from_csv
-
+            csv_filepath, netcdf_filepath = (
+                self.current_simulation.precipitation.csv_filepath,
+                self.current_simulation.precipitation.netcdf_filepath,
+            )
+            netcdf_global, netcdf_raster = (
+                self.current_simulation.precipitation.netcdf_global,
+                self.current_simulation.precipitation.netcdf_raster,
+            )
             if precipitation_type == EventTypes.CONSTANT.value:
                 self.tc.create_simulation_constant_precipitation(
                     sim_id, value=values, units=units, duration=duration, offset=offset
                 )
-            elif precipitation_type == EventTypes.CUSTOM.value:
-                if from_csv:
-                    for values_chunk in split_to_even_chunks(values, 300):
-                        chunk_offset = values_chunk[0][0]
-                        values_chunk = [[t - chunk_offset, v] for t, v in values_chunk]
-                        self.tc.create_simulation_custom_precipitation(
-                            sim_id,
-                            values=values_chunk,
-                            units=units,
-                            duration=duration,
-                            offset=offset + chunk_offset,
-                            interpolate=interpolate,
-                        )
+            elif precipitation_type == EventTypes.FROM_CSV.value:
+                for values_chunk in split_to_even_chunks(values, 300):
+                    chunk_offset = values_chunk[0][0]
+                    values_chunk = [[t - chunk_offset, v] for t, v in values_chunk]
+                    self.tc.create_simulation_custom_precipitation(
+                        sim_id,
+                        values=values_chunk,
+                        units=units,
+                        duration=duration,
+                        offset=offset + chunk_offset,
+                        interpolate=interpolate,
+                    )
+            elif precipitation_type == EventTypes.FROM_NETCDF.value:
+                filename = os.path.basename(netcdf_filepath)
+                if netcdf_global:
+                    upload = self.tc.create_simulation_global_netcdf_precipitation(sim_id, filename=filename)
                 else:
-                    filename = os.path.basename(filepath)
-                    upload = self.tc.create_simulation_custom_netcdf_precipitation(sim_id, filename=filename)
-                    upload_local_file(upload, filepath)
+                    upload = self.tc.create_simulation_raster_netcdf_precipitation(sim_id, filename=filename)
+                upload_local_file(upload, netcdf_filepath)
             elif precipitation_type == EventTypes.DESIGN.value:
                 self.tc.create_simulation_custom_precipitation(
                     sim_id, values=values, units=units, duration=duration, offset=offset
@@ -975,7 +986,7 @@ class SimulationRunner(QRunnable):
                     speed_value=speed,
                     direction_value=direction,
                 )
-            elif wind_type == EventTypes.CUSTOM.value:
+            elif wind_type == EventTypes.FROM_CSV.value:
                 self.tc.create_simulation_custom_wind(
                     sim_id,
                     offset=offset,