+++ HISTORY.rst
@@ -4,7 +4,7 @@ History
 3.9.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Updated threedi-schema version to 0.219.1
 
 
 3.9.1 (2024-03-14)