+++ HISTORY.rst
@@ -5,13 +5,12 @@ History
 ------------------
 
 - Fixes/enhancements: #527
-
+- Fixed widget focus for the simulation model selection window.
 
 3.9.1 (2024-03-14)
 ------------------
 
-- Fix logout error: #533
-
+- Fixes/enhancements: #528, #533
 
 3.9 (2024-01-16)
 ----------------