+++ HISTORY.rst
@@ -4,8 +4,8 @@ History
 3.9.2 (unreleased)
 ------------------
 
+- Fixes/enhancements: #527
 - Fixed widget focus for the simulation model selection window.
-
 
 3.9.1 (2024-03-14)
 ------------------